IntelMap/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
pydeck>=0.8.0
humanize>=4.6.0
geopy>=2.3.0
spacy>=3.7.0
telethon>=1.28.5
python-dotenv>=1.0.0
python-dateutil>=2.8.2